refactor(local): hide local implementation from constraints model
